move function and add docstring
