+++ src/napari/_app_model/utils.py
@@ -5,9 +5,6 @@ from app_model.types import Action, MenuItem, SubmenuItem
 
 from napari._app_model import get_app_model
 from napari._app_model.constants import MenuGroup, MenuId
-from napari.settings import get_settings
-from napari.utils.action_manager import action_manager
-from napari.utils.interactions import Shortcut
 
 MenuOrSubmenu = Union[MenuItem, SubmenuItem]
 
@@ -128,28 +125,3 @@ def get_dummy_action(menu_id: MenuId) -> tuple[Action, str]:
         palette=False,
     )
     return action, context_key
-
-
-def get_command_shortcut_and_description(
-    command_id: str,
-) -> tuple[str | None, str | None]:
-    """Get the command shortcut and description from a command id."""
-    app = get_app_model()
-
-    app_model_keybinding = app.keybindings.get_keybinding(command_id)
-    if app_model_keybinding is not None:
-        return (
-            Shortcut(app_model_keybinding.keybinding).platform,
-            app.commands[command_id].title,
-        )
-
-    # might be an action_manager action
-    settings_shortcuts = get_settings().shortcuts.shortcuts
-    action_manager_keybinding = settings_shortcuts.get(command_id, [None])[0]
-    if action_manager_keybinding is not None:
-        return (
-            Shortcut(action_manager_keybinding).platform,
-            action_manager._actions[command_id].description,
-        )
-
-    return None, None
